rename(route-folder): main 페이지 상수 컨벤션 적용
