--- src/app/(home)/_components/DefaultSheetContent/_internal/PoliceSection/POLICE_ITEMS.ts
@@ -1,12 +1,0 @@
-export const POLICE_ITEMS = [
-  {
-    href: "https://www.finditem.kr/public-data?type=lost",
-    headLabel: "분실",
-    label: "했어요",
-  },
-  {
-    href: "https://www.finditem.kr/public-data?type=found",
-    headLabel: "발견",
-    label: "했어요",
-  },
-];
