--- src/app/(home)/_components/DefaultSheetContent/_internal/SupportLinkSection/SUPPORT_MENU_ITEMS.ts
@@ -1,10 +1,0 @@
-export const SUPPORT_MENU_ITEMS = [
-  {
-    label: "키워드 알람 설정하기",
-    href: "/mypage/notifications",
-  },
-  {
-    label: "공지사항",
-    href: "/notice",
-  },
-];
