+++ src/app/(home)/_components/HomeFilterSection/_internal/CategoryFilter/CategoryFilter.tsx
@@ -4,7 +4,7 @@ import { useRef, useState } from "react";
 import { createPortal } from "react-dom";
 import { Filter } from "@/components";
 import { MAP_CATEGORY_FILTER_OPTIONS } from "@/constants";
-import { CATEGORY_FILTER_DROPDOWN_MIN_WIDTH_PX } from "../../../../_constants/FILTER";
+import { CATEGORY_FILTER_DROPDOWN_MIN_WIDTH_PX } from "../../../HOME_CONST";
 import { cn } from "@/utils";
 import { usePopoverOutsideClose, usePopoverPosition } from "@/hooks";
