+++ src/app/(home)/_components/MainKakaoMap/MainKakaoMap.tsx
@@ -4,7 +4,7 @@ import { BaseKakaoMap } from "@/components";
 import { useMainKakaoMap } from "../../_hooks";
 import { useGetMarker, isMarkerFetchDisabledByZoom } from "@/api/fetch/mapController";
 import { useRouter, useSearchParams } from "next/navigation";
-import { MARKER_ID } from "../../_constants/QUERY_PARAMS";
+import { MARKER_ID } from "../HOME_CONST";
 import { useMainKakaoMapStore } from "@/store";
 
 const MainKakaoMap = () => {
