+++ src/app/(home)/_components/MainSearchEmpty/MainSearchEmpty.tsx
@@ -2,7 +2,7 @@ import { Button, Icon } from "@/components";
 import { cn } from "@/utils";
 import Image from "next/image";
 import Link from "next/link";
-import { BUTTON_DEFAULT_STYLE, WRITE_BUTTONS } from "./WRITE_BUTTONS";
+import { BUTTON_DEFAULT_STYLE, WRITE_BUTTONS } from "../HOME_CONST";
 
 const MainSearchEmpty = () => {
   return (
