+++ src/app/(home)/_components/PermissionBottomSheet/PermissionBottomSheet.tsx
@@ -4,7 +4,7 @@ import { useToast } from "@/context/ToastContext";
 import { useMainKakaoMapStore } from "@/store";
 import { clearMainGeoSessionConfirmed, markMainGeoSessionConfirmed } from "@/utils/mainGeoSession";
 import { useState } from "react";
-import { PERMISSION_CONFIG, PERMISSION_ITEM } from "../../_constants/PERMISSION_CONFIG";
+import { PERMISSION_CONFIG, PERMISSION_ITEM } from "../HOME_CONST";
 import { syncWebPushSubscription } from "@/utils";
 
 interface DetailPermissionSheetProps {
