--- src/app/(home)/_constants/HEIGHT_PX.ts
@@ -1,10 +1,0 @@
-export const BOTTOM_OFFSET_PX = 86.67;
-export const HEADER_HEIGHT_PX = 85;
-export const MIN_HEIGHT_PX = 27;
-export const INITIAL_HEIGHT_PX = 591;
-export const SNAP_RATIOS = [0, 0.3, 0.5, 0.75, 1] as const;
-
-/** 바텀시트 핸들 영역 높이 (pt-3 + bar 3px + pb-5) */
-export const SHEET_HANDLE_HEIGHT_PX = 35;
-/** 스크롤 콘텐츠 하단 패딩 (pb-[18px]) */
-export const SHEET_CONTENT_BOTTOM_PADDING_PX = 18;
