--- src/app/(home)/_constants/QUERY_PARAMS.ts
@@ -1,4 +1,0 @@
-export const MARKER_ID = "marker-id" as const;
-
-export const POST_TYPE = "post-type" as const;
-export const CATEGORY = "category" as const;
