+++ src/app/(home)/_hooks/useBottomSheetHeight/useBottomSheetHeight.ts
@@ -3,8 +3,7 @@
 import { animate, useMotionValue, useMotionValueEvent } from "framer-motion";
 import { useEffect, useRef, useState } from "react";
 import { useSearchParams } from "next/navigation";
-import { INITIAL_HEIGHT_PX, MIN_HEIGHT_PX } from "../../_constants/HEIGHT_PX";
-import { MARKER_ID } from "../../_constants/QUERY_PARAMS";
+import { INITIAL_HEIGHT_PX, MIN_HEIGHT_PX, MARKER_ID } from "../../_components/HOME_CONST";
 import {
   getMaxHeightPx,
   getSnapHeights,
