+++ src/app/(home)/_hooks/useHomeFilterQuery/useHomeFilterQuery.ts
@@ -2,7 +2,7 @@
 
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
 import { CATEGORY_OPTIONS } from "@/constants";
-import { CATEGORY, POST_TYPE } from "../../_constants/QUERY_PARAMS";
+import { CATEGORY, POST_TYPE } from "../../_components/HOME_CONST";
 import { PostFilterChipValue } from "../../_types/PostFilterChipValue";
 
 const getSelectedPostFilterFromQuery = (postType: string | null): PostFilterChipValue => {
